✨ feat(metrics): Prompt for post story points
- Refactors the logic for prompting users to re-estimate post story points for completed issues.
- Extracts the post story point prompting logic into a reusable utility function.
- Integrates the new utility function into both the guided workflow and interactive quit actions.
- Updates the `.luma_metrics.json` to reflect completed issues with actual completion dates and mandays for two previously in-progress issues.
