+++ .luma_metrics.json
@@ -115,20 +115,20 @@
       "issue_url": "https://github.com/oatrice/Luma/issues/32",
       "repository": "oatrice/Luma",
       "project_name": "Luma",
-      "issue_status": "🟡 In Progress",
+      "issue_status": "✅ Complete",
       "estimate_points": 2,
       "post_story_point": null,
       "estimated_mandays": 1.0,
       "start_datetime": "2026-04-02T08:36:34",
-      "actual_mandays": 0.0,
+      "actual_mandays": 0.5,
       "due_date": "2026-04-03T23:59:59",
-      "actual_completion_date": null,
+      "actual_completion_date": "2026-04-02T11:52:57",
       "effort_level": "Low",
       "notes": null,
-      "gh_closed_at": null,
-      "gh_mandays": null,
+      "gh_closed_at": "2026-04-02T11:52:57Z",
+      "gh_mandays": 0.5,
       "created_at": "2026-04-02T08:25:00Z",
-      "updated_at": "2026-04-02T17:06:56.468764"
+      "updated_at": "2026-04-02T18:54:11.211993"
     },
     "oatrice/Luma#33": {
       "issue_key": "oatrice/Luma#33",
@@ -137,20 +137,20 @@
       "issue_url": "https://github.com/oatrice/Luma/issues/33",
       "repository": "oatrice/Luma",
       "project_name": "Luma",
-      "issue_status": "🟡 In Progress",
+      "issue_status": "✅ Complete",
       "estimate_points": 2,
       "post_story_point": null,
       "estimated_mandays": 1.0,
       "start_datetime": "2026-04-02T08:36:36",
-      "actual_mandays": 0.0,
+      "actual_mandays": 0.5,
       "due_date": "2026-04-03T23:59:59",
-      "actual_completion_date": null,
+      "actual_completion_date": "2026-04-02T11:52:58",
       "effort_level": "Low",
       "notes": null,
-      "gh_closed_at": null,
-      "gh_mandays": null,
+      "gh_closed_at": "2026-04-02T11:52:58Z",
+      "gh_mandays": 0.5,
       "created_at": "2026-04-02T08:25:21Z",
-      "updated_at": "2026-04-02T17:06:59.723977"
+      "updated_at": "2026-04-02T18:54:12.079204"
     }
   }
 }
